Add capability-gated environment dashboard and MQTT status queries
