+++ docs/web-ui.md
@@ -207,7 +207,7 @@ The Settings collapsible group lives in the shared sidebar `NAV` (`system/web/sr
 | Plugins | `/setting#plugins` |
 | Timezone | `/setting#timezone` |
 
-Monitor leaves serialize as the plain id, e.g. `/monitor#overview`, `/monitor#pairing`, `/monitor#system`, `/monitor#flow`. Defaults: `/monitor` with no/invalid hash → `overview`; `/setting` with no/invalid hash → `general` (URL normalized to `/setting#general`). Deep-links (e.g. `/setting#wifi`) and browser back/forward are honored via a `useLocation`-driven effect. Non-debug users only see the leaves in `PUBLIC_SECTIONS` (which includes Chat, Overview, **Pairing**, Info, Flow, Camera, Users, **Logs**, **CLI**, and the public Settings leaves General/Wi-Fi/My Voice/Face/MCP Tools/Plugins/Timezone); Bluetooth remains available by direct URL but is hidden from navigation. `?debug=true` reveals the rest (Sensing, Analytics, Servo, API Docs, Agent gateway, and the deeper Settings leaves AI Brain/Runtime/Language/Voice/Realtime/Channels/MQTT). Pressing `update` swaps the button for `updating…` immediately — the button never says "OK", which would read as "done" for a request that has only STARTED the install (and, for a component that finishes in seconds, arrived before the row could even show progress). Failures show the server's own reason (`rate-limited, retry in 8s`, `bootstrap unreachable`) rather than a bare "Failed". While an install runs, that row shows `updating…` in place of the button (an install takes tens of seconds — the component stops, is rebuilt and restarts — and a row that just sits there invites a second click, which is how a device once lost its HAL runtime). The `update` buttons in the Overview **Versions** card (Web / OS / HAL / Agent rows, plus Bootstrap and Device in debug) are gated the same way — regular viewers get no one-click OTA trigger. The top-bar **Debug** toggle beside the Dark/Light button toggles that query parameter while preserving the active route hash and any other query parameters; its amber state indicates that debug mode is enabled.
+Monitor leaves serialize as the plain id, e.g. `/monitor#overview`, `/monitor#pairing`, `/monitor#system`, `/monitor#flow`. Defaults: `/monitor` with no/invalid hash → `overview`; `/setting` with no/invalid hash → `general` (URL normalized to `/setting#general`). Deep-links (e.g. `/setting#wifi`) and browser back/forward are honored via a `useLocation`-driven effect. Non-debug users only see the leaves in `PUBLIC_SECTIONS` (which includes Chat, Overview, **Pairing**, Info, Flow, Camera, **Sensing**, Users, **Logs**, **CLI**, and the public Settings leaves General/Wi-Fi/My Voice/Face/MCP Tools/Plugins/Timezone); Bluetooth remains available by direct URL but is hidden from navigation. `?debug=true` reveals the rest (Analytics, Servo, API Docs, Agent gateway, and the deeper Settings leaves AI Brain/Runtime/Language/Voice/Realtime/Channels/MQTT). Pressing `update` swaps the button for `updating…` immediately — the button never says "OK", which would read as "done" for a request that has only STARTED the install (and, for a component that finishes in seconds, arrived before the row could even show progress). Failures show the server's own reason (`rate-limited, retry in 8s`, `bootstrap unreachable`) rather than a bare "Failed". While an install runs, that row shows `updating…` in place of the button (an install takes tens of seconds — the component stops, is rebuilt and restarts — and a row that just sits there invites a second click, which is how a device once lost its HAL runtime). The `update` buttons in the Overview **Versions** card (Web / OS / HAL / Agent rows, plus Bootstrap and Device in debug) are gated the same way — regular viewers get no one-click OTA trigger. The top-bar **Debug** toggle beside the Dark/Light button toggles that query parameter while preserving the active route hash and any other query parameters; its amber state indicates that debug mode is enabled.
 
 The Overview **Versions** card has a fifth action column with `restart` for OS Server and HAL, including outside debug mode. Each button calls the admin-protected `POST /api/system/restart/:target` (`os-server` or `hal`). The server schedules the restart after 2 seconds and returns HTTP 202. The button shows `queued` and disables repeat clicks for 15 seconds; this acknowledges scheduling, not service recovery. Existing monitor polling refreshes status/uptime after reconnection. Errors remain visible beside the button. Restart is disabled while the row is known to be updating; OTA activity is polled in normal mode too. Narrow cards scroll horizontally to keep all five columns accessible.
 
@@ -719,6 +719,34 @@ Chat UI → POST /api/sensing/event → SensingHandler
 
 ---
 
+### 5.8 Device → Sensing
+
+The Sensing navigation entry is available without debug mode when the device declares `vision` or `environment` in
+`GET /api/system/info` → `capabilities`. Camera sensing cards require `vision`.
+The read-only **Environment · SEN55** card requires an explicit `environment`
+capability; it is hidden and sends no requests while capabilities are loading
+or when that capability is absent.
+
+While mounted, the environment card reads `GET /api/hardware/environment/status`
+every 3 seconds through the existing authenticated OS hardware reverse proxy to
+HAL `GET /environment/status`. This browser refresh interval is independent of
+HAL's configurable `poll_interval_s`; it does not change acquisition frequency.
+No custom OS environment endpoint or OS → agent event is added.
+
+The card shows sensor state, sample timestamp, stale status, errors,
+and eight measurements: temperature (°C), humidity (%), PM1 / PM2.5 / PM4 /
+PM10 (µg/m³), VOC index, and NOx index. Unavailable values appear as `—`, never
+zero. Stale measurements are also replaced with `—`; a request failure is shown
+as an error so previous readings cannot be mistaken for live data. No good/bad air
+quality labels, thresholds, or alerts are assigned. A collapsed technical
+section exposes I2C bus, sensor status register, and HAL polling/retry/staleness/
+recovery timings from `status.timing`.
+
+Lamp still ships with `environment` commented out in `ROBOT.md` and SEN55
+disabled in `sen55.json`, so this card remains hidden until the capability is
+declared. See [Lamp environmental sensing](../robots/lamp/docs/environment-sensing.md)
+for wiring, enabling, and the HAL data contract.
+
 ## 6. LED Color API
 
 ### Problem
@@ -812,3 +840,14 @@ path instead — `make upload-hal` then `make promote-hal`, which versions the
 artifact and rolls it out.
 
 Harness final delivery records `harness_response` in flow JSONL with the original device run ID and complete `text`. Web Chat uses this event to recover pending results after SSE disconnects or page reloads. Live delivery still emits `chat_response` with state `final`.
+
+### Sensing component structure
+
+`monitor/SensingSection.tsx` only composes capability-gated sections. Components
+live under `monitor/sensing/`: one file per card, with shared `CardHeader`, types
+and formatters. `useVisionSensing` polls once for all vision cards;
+`useEnvironment` independently polls SEN55. The `visionApi.ts` and
+`environmentApi.ts` clients own OS-server requests, response checks and errors;
+cards do not fetch directly. Both clients use the existing authenticated
+`/api/hardware/*` proxy. Vision HTTP/response failures show an error instead of
+leaving the loading placeholder or old readings on screen.
